[#3629] Refactor filtered server map creation
