+++ web/src/main/java/com/navercorp/pinpoint/web/service/FilteredMapServiceImpl.java
@@ -20,9 +20,6 @@ import com.google.common.collect.Lists;
 import com.navercorp.pinpoint.common.server.bo.SpanBo;
 import com.navercorp.pinpoint.common.server.bo.SpanEventBo;
 import com.navercorp.pinpoint.common.service.ServiceTypeRegistryService;
-import com.navercorp.pinpoint.common.trace.HistogramSchema;
-import com.navercorp.pinpoint.common.trace.HistogramSlot;
-import com.navercorp.pinpoint.common.trace.ServiceType;
 import com.navercorp.pinpoint.common.util.TransactionId;
 import com.navercorp.pinpoint.web.applicationmap.ApplicationMap;
 import com.navercorp.pinpoint.web.applicationmap.ApplicationMapBuilder;
@@ -31,29 +28,26 @@ import com.navercorp.pinpoint.web.applicationmap.ApplicationMapWithScatterData;
 import com.navercorp.pinpoint.web.applicationmap.ApplicationMapWithScatterScanResult;
 import com.navercorp.pinpoint.web.applicationmap.appender.histogram.DefaultNodeHistogramFactory;
 import com.navercorp.pinpoint.web.applicationmap.appender.histogram.NodeHistogramFactory;
-import com.navercorp.pinpoint.web.applicationmap.appender.histogram.datasource.ResponseHistogramBuilderNodeHistogramDataSource;
+import com.navercorp.pinpoint.web.applicationmap.appender.histogram.datasource.ResponseHistogramsNodeHistogramDataSource;
 import com.navercorp.pinpoint.web.applicationmap.appender.histogram.datasource.WasNodeHistogramDataSource;
 import com.navercorp.pinpoint.web.applicationmap.appender.server.DefaultServerInstanceListFactory;
 import com.navercorp.pinpoint.web.applicationmap.appender.server.ServerInstanceListFactory;
 import com.navercorp.pinpoint.web.applicationmap.appender.server.datasource.AgentInfoServerInstanceListDataSource;
 import com.navercorp.pinpoint.web.applicationmap.appender.server.datasource.ServerInstanceListDataSource;
 import com.navercorp.pinpoint.web.applicationmap.link.LinkType;
-import com.navercorp.pinpoint.web.applicationmap.rawdata.LinkDataDuplexMap;
-import com.navercorp.pinpoint.web.applicationmap.rawdata.LinkDataMap;
 import com.navercorp.pinpoint.web.dao.ApplicationTraceIndexDao;
 import com.navercorp.pinpoint.web.dao.TraceDao;
 import com.navercorp.pinpoint.web.filter.Filter;
+import com.navercorp.pinpoint.web.scatter.ScatterData;
 import com.navercorp.pinpoint.web.security.ServerMapDataFilter;
-import com.navercorp.pinpoint.web.util.TimeWindow;
-import com.navercorp.pinpoint.web.util.TimeWindowDownSampler;
+import com.navercorp.pinpoint.web.service.map.FilteredMap;
+import com.navercorp.pinpoint.web.service.map.FilteredMapBuilder;
 import com.navercorp.pinpoint.web.vo.Application;
 import com.navercorp.pinpoint.web.vo.LimitedScanResult;
 import com.navercorp.pinpoint.web.vo.LoadFactor;
 import com.navercorp.pinpoint.web.vo.Range;
-import com.navercorp.pinpoint.web.vo.ResponseHistogramBuilder;
 import com.navercorp.pinpoint.web.vo.SelectedScatterArea;
-import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.lang3.StringUtils;
+import com.navercorp.pinpoint.web.vo.scatter.ApplicationScatterScanResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -62,7 +56,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StopWatch;
 
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -213,11 +207,18 @@ public class FilteredMapServiceImpl implements FilteredMapService {
         if (transactionId == null) {
             throw new NullPointerException("transactionId must not be null");
         }
-        List<TransactionId> transactionIdList = new ArrayList<>();
-        transactionIdList.add(transactionId);
+        List<TransactionId> transactionIdList = Collections.singletonList(transactionId);
         // FIXME from,to -1
         Range range = new Range(-1, -1);
-        return selectApplicationMap(transactionIdList, range, range, Filter.NONE, version);
+
+        final List<List<SpanBo>> filterList = selectFilteredSpan(transactionIdList, Filter.NONE);
+        FilteredMapBuilder filteredMapBuilder = new FilteredMapBuilder(applicationFactory, registry, range, version);
+        filteredMapBuilder.serverMapDataFilter(serverMapDataFilter);
+        filteredMapBuilder.addTransactions(filterList);
+        FilteredMap filteredMap = filteredMapBuilder.build();
+
+        ApplicationMap map = createMap(range, filteredMap);
+        return map;
     }
 
     /**
@@ -236,11 +237,15 @@ public class FilteredMapServiceImpl implements FilteredMapService {
         watch.start();
 
         final List<List<SpanBo>> filterList = selectFilteredSpan(transactionIdList, filter);
+        FilteredMapBuilder filteredMapBuilder = new FilteredMapBuilder(applicationFactory, registry, originalRange, version);
+        filteredMapBuilder.serverMapDataFilter(serverMapDataFilter);
+        filteredMapBuilder.addTransactions(filterList);
+        FilteredMap filteredMap = filteredMapBuilder.build();
 
-        DotExtractor dotExtractor = createDotExtractor(scanRange, filterList);
-        ApplicationMap map = createMap(originalRange, scanRange, filterList, version);
+        ApplicationMap map = createMap(originalRange, filteredMap);
 
-        ApplicationMapWithScatterScanResult applicationMapWithScatterScanResult = new ApplicationMapWithScatterScanResult(map, dotExtractor.getApplicationScatterScanResult());
+        List<ApplicationScatterScanResult> applicationScatterScanResult = filteredMap.getApplicationScatterScanResult(scanRange.getFrom(), scanRange.getTo());
+        ApplicationMapWithScatterScanResult applicationMapWithScatterScanResult = new ApplicationMapWithScatterScanResult(map, applicationScatterScanResult);
 
         watch.stop();
         logger.debug("Select filtered application map elapsed. {}ms", watch.getTotalTimeMillis());
@@ -261,11 +266,15 @@ public class FilteredMapServiceImpl implements FilteredMapService {
         watch.start();
 
         final List<List<SpanBo>> filterList = selectFilteredSpan(transactionIdList, filter);
+        FilteredMapBuilder filteredMapBuilder = new FilteredMapBuilder(applicationFactory, registry, originalRange, version);
+        filteredMapBuilder.serverMapDataFilter(serverMapDataFilter);
+        filteredMapBuilder.addTransactions(filterList);
+        FilteredMap filteredMap = filteredMapBuilder.build();
 
-        DotExtractor dotExtractor = createDotExtractor(scanRange, filterList);
-        ApplicationMap map = createMap(originalRange, scanRange, filterList, version);
+        ApplicationMap map = createMap(originalRange, filteredMap);
 
-        ApplicationMapWithScatterData applicationMapWithScatterData = new ApplicationMapWithScatterData(map, dotExtractor.getApplicationScatterData(originalRange.getFrom(), originalRange.getTo(), xGroupUnit, yGroupUnit));
+        Map<Application, ScatterData> applicationScatterData = filteredMap.getApplicationScatterData(originalRange.getFrom(), originalRange.getTo(), xGroupUnit, yGroupUnit);
+        ApplicationMapWithScatterData applicationMapWithScatterData = new ApplicationMapWithScatterData(map, applicationScatterData);
 
         watch.stop();
         logger.debug("Select filtered application map elapsed. {}ms", watch.getTotalTimeMillis());
@@ -284,89 +293,8 @@ public class FilteredMapServiceImpl implements FilteredMapService {
         return filterList2(originalList, filter);
     }
 
-    private DotExtractor createDotExtractor(Range scanRange, List<List<SpanBo>> filterList) {
-        final DotExtractor dotExtractor = new DotExtractor(scanRange, applicationFactory);
-
-        for (List<SpanBo> transaction : filterList) {
-            for (SpanBo span : transaction) {
-                final Application spanApplication = this.applicationFactory.createApplication(span.getApplicationId(), span.getApplicationServiceType());
-                if (!spanApplication.getServiceType().isRecordStatistics() || spanApplication.getServiceType().isRpcClient()) {
-                    continue;
-                }
-
-                dotExtractor.addDot(span);
-            }
-        }
-
-        return dotExtractor;
-    }
-
-    private ApplicationMap createMap(Range range, Range scanRange, List<List<SpanBo>> filterList, int version) {
-        // TODO inject TimeWindow from elsewhere
-        final TimeWindow window = new TimeWindow(range, TimeWindowDownSampler.SAMPLER);
-
-
-        final LinkDataDuplexMap linkDataDuplexMap = new LinkDataDuplexMap();
-
-        final ResponseHistogramBuilder mapHistogramSummary = new ResponseHistogramBuilder(range);
-        /*
-         * Convert to statistical data
-         */
-        for (List<SpanBo> transaction : filterList) {
-            final Map<Long, SpanBo> transactionSpanMap = checkDuplicatedSpanId(transaction);
-
-            for (SpanBo span : transaction) {
-                final Application parentApplication = createParentApplication(span, transactionSpanMap, version);
-                final Application spanApplication = this.applicationFactory.createApplication(span.getApplicationId(), span.getApplicationServiceType());
-
-                // records the Span's response time statistics
-                recordSpanResponseTime(spanApplication, span, mapHistogramSummary, span.getCollectorAcceptTime());
-
-                if (!spanApplication.getServiceType().isRecordStatistics() || spanApplication.getServiceType().isRpcClient()) {
-                    // span's serviceType is probably not set correctly
-                    logger.warn("invalid span application:{}", spanApplication);
-                    continue;
-                }
-
-                final short slotTime = getHistogramSlotTime(span, spanApplication.getServiceType());
-                // might need to reconsider using collector's accept time for link statistics.
-                // we need to convert to time window's timestamp. If not, it may lead to OOM due to mismatch in timeslots. 
-                long timestamp = window.refineTimestamp(span.getCollectorAcceptTime());
-
-                if (parentApplication.getServiceType() == ServiceType.USER) {
-                    // Outbound data
-                    if (logger.isTraceEnabled()) {
-                        logger.trace("span user:{} {} -> span:{} {}", parentApplication, span.getAgentId(), spanApplication, span.getAgentId());
-                    }
-                    final LinkDataMap sourceLinkData = linkDataDuplexMap.getSourceLinkDataMap();
-                    sourceLinkData.addLinkData(parentApplication, span.getAgentId(), spanApplication,  span.getAgentId(), timestamp, slotTime, 1);
-
-                    if (logger.isTraceEnabled()) {
-                        logger.trace("span target user:{} {} -> span:{} {}", parentApplication, span.getAgentId(), spanApplication, span.getAgentId());
-                    }
-                    // Inbound data
-                    final LinkDataMap targetLinkDataMap = linkDataDuplexMap.getTargetLinkDataMap();
-                    targetLinkDataMap.addLinkData(parentApplication, span.getAgentId(), spanApplication, span.getAgentId(), timestamp, slotTime, 1);
-                } else {
-                    // Inbound data
-                    if (logger.isTraceEnabled()) {
-                        logger.trace("span target parent:{} {} -> span:{} {}", parentApplication, span.getAgentId(), spanApplication, span.getAgentId());
-                    }
-                    final LinkDataMap targetLinkDataMap = linkDataDuplexMap.getTargetLinkDataMap();
-                    targetLinkDataMap.addLinkData(parentApplication, span.getAgentId(), spanApplication, span.getAgentId(), timestamp, slotTime, 1);
-                }
-
-                if (serverMapDataFilter != null && serverMapDataFilter.filter(spanApplication)) {
-                    continue;
-                }
-                
-                addNodeFromSpanEvent(span, window, linkDataDuplexMap, transactionSpanMap);
-            }
-        }
-
-        mapHistogramSummary.build();
-
-        WasNodeHistogramDataSource wasNodeHistogramDataSource = new ResponseHistogramBuilderNodeHistogramDataSource(mapHistogramSummary);
+    private ApplicationMap createMap(Range range, FilteredMap filteredMap) {
+        WasNodeHistogramDataSource wasNodeHistogramDataSource = new ResponseHistogramsNodeHistogramDataSource(filteredMap.getResponseHistograms());
         NodeHistogramFactory nodeHistogramFactory = new DefaultNodeHistogramFactory(wasNodeHistogramDataSource);
 
         ServerInstanceListDataSource serverInstanceListDataSource = new AgentInfoServerInstanceListDataSource(agentInfoService);
@@ -376,133 +304,13 @@ public class FilteredMapServiceImpl implements FilteredMapService {
         applicationMapBuilder.linkType(LinkType.DETAILED);
         applicationMapBuilder.includeNodeHistogram(nodeHistogramFactory);
         applicationMapBuilder.includeServerInfo(serverInstanceListFactory);
-        ApplicationMap map = applicationMapBuilder.build(linkDataDuplexMap);
+        ApplicationMap map = applicationMapBuilder.build(filteredMap.getLinkDataDuplexMap());
 
         if(serverMapDataFilter != null) {
             map = serverMapDataFilter.dataFiltering(map);
         }
-        
+
         return map;
-    }
-
-    private Map<Long, SpanBo> checkDuplicatedSpanId(List<SpanBo> transaction) {
-        final Map<Long, SpanBo> transactionSpanMap = new HashMap<>();
-        for (SpanBo span : transaction) {
-            final SpanBo old = transactionSpanMap.put(span.getSpanId(), span);
-            if (old != null) {
-                logger.warn("duplicated span found:{}", old);
-            }
-        }
-        return transactionSpanMap;
-    }
-
-    private void recordSpanResponseTime(Application application, SpanBo span, ResponseHistogramBuilder responseHistogramBuilder, long timeStamp) {
-        responseHistogramBuilder.addHistogram(application, span, timeStamp);
-    }
-
-    private void addNodeFromSpanEvent(SpanBo span, TimeWindow window, LinkDataDuplexMap linkDataDuplexMap, Map<Long, SpanBo> transactionSpanMap) {
-        /*
-         * add span event statistics
-         */
-        final List<SpanEventBo> spanEventBoList = span.getSpanEventBoList();
-        if (CollectionUtils.isEmpty(spanEventBoList)) {
-            return;
-        }
-        final Application srcApplication = applicationFactory.createApplication(span.getApplicationId(), span.getApplicationServiceType());
-
-        LinkDataMap sourceLinkDataMap = linkDataDuplexMap.getSourceLinkDataMap();
-        for (SpanEventBo spanEvent : spanEventBoList) {
-
-            ServiceType destServiceType = registry.findServiceType(spanEvent.getServiceType());
-            if (!destServiceType.isRecordStatistics()) {
-                // internal method
-                continue;
-            }
-            // convert to Unknown if destServiceType is a rpc client and there is no acceptor.
-            // acceptor exists if there is a span with spanId identical to the current spanEvent's next spanId.
-            // logic for checking acceptor
-            if (destServiceType.isRpcClient()) {
-                if (!transactionSpanMap.containsKey(spanEvent.getNextSpanId())) {
-                    destServiceType = ServiceType.UNKNOWN;
-                }
-            }
-
-            String dest = spanEvent.getDestinationId();
-            if (dest == null) {
-                dest = "Unknown";
-            }
-            
-            final Application destApplication = this.applicationFactory.createApplication(dest, destServiceType);
-
-            final short slotTime = getHistogramSlotTime(spanEvent, destServiceType);
-
-            // FIXME
-            final long spanEventTimeStamp = window.refineTimestamp(span.getStartTime() + spanEvent.getStartElapsed());
-            if (logger.isTraceEnabled()) {
-                logger.trace("spanEvent  src:{} {} -> dest:{} {}", srcApplication, span.getAgentId(), destApplication, spanEvent.getEndPoint());
-            }
-            // endPoint may be null
-            final String destinationAgentId = StringUtils.defaultString(spanEvent.getEndPoint());
-            sourceLinkDataMap.addLinkData(srcApplication, span.getAgentId(), destApplication, destinationAgentId, spanEventTimeStamp, slotTime, 1);
-        }
-    }
-
-    private Application createParentApplication(SpanBo span, Map<Long, SpanBo> transactionSpanMap, int version) {
-        final SpanBo parentSpan = transactionSpanMap.get(span.getParentSpanId());
-        if (span.isRoot() || parentSpan == null) {
-            ServiceType spanServiceType = this.registry.findServiceType(span.getServiceType());
-            if (spanServiceType.isQueue()) {
-                String applicationName = span.getAcceptorHost();
-                ServiceType serviceType = spanServiceType;
-                return this.applicationFactory.createApplication(applicationName, serviceType);
-            } else {
-                String applicationName;
-                // FIXME magic number, remove after front end UI changes and simply use the newer one
-                if (version >= 4) {
-                    ServiceType applicationServiceType = this.registry.findServiceType(span.getApplicationServiceType());
-                    applicationName = span.getApplicationId() + "_" + applicationServiceType;
-                } else {
-                    applicationName = span.getApplicationId();
-                }
-                ServiceType serviceType = ServiceType.USER;
-                return this.applicationFactory.createApplication(applicationName, serviceType);
-            }
-        } else {
-            // create virtual queue node if current' span's service type is a queue AND :
-            // 1. parent node's application service type is not a queue (it may have come from a queue that is traced)
-            // 2. current node's application service type is not a queue (current node may be a queue that is traced)
-            ServiceType spanServiceType = this.registry.findServiceType(span.getServiceType());
-            if (spanServiceType.isQueue()) {
-                ServiceType parentApplicationServiceType = this.registry.findServiceType(parentSpan.getApplicationServiceType());
-                ServiceType spanApplicationServiceType = this.registry.findServiceType(span.getApplicationServiceType());
-                if (!parentApplicationServiceType.isQueue() && !spanApplicationServiceType.isQueue()) {
-                    String parentApplicationName = span.getAcceptorHost();
-                    if (parentApplicationName == null) {
-                        parentApplicationName = span.getRemoteAddr();
-                    }
-                    short parentServiceType = span.getServiceType();
-                    return this.applicationFactory.createApplication(parentApplicationName, parentServiceType);
-                }
-            }
-            String parentApplicationName = parentSpan.getApplicationId();
-            short parentServiceType = parentSpan.getApplicationServiceType();
-            return this.applicationFactory.createApplication(parentApplicationName, parentServiceType);
-        }
-    }
-
-    private short getHistogramSlotTime(SpanEventBo spanEvent, ServiceType serviceType) {
-        return getHistogramSlotTime(spanEvent.hasException(), spanEvent.getEndElapsed(), serviceType);
-    }
-
-    private short getHistogramSlotTime(SpanBo span, ServiceType serviceType) {
-        boolean allException = span.getErrCode() != 0;
-        return getHistogramSlotTime(allException, span.getElapsed(), serviceType);
-    }
-
-    private short getHistogramSlotTime(boolean hasException, int elapsedTime, ServiceType serviceType) {
-        final HistogramSchema schema = serviceType.getHistogramSchema();
-        final HistogramSlot histogramSlot = schema.findHistogramSlot(elapsedTime, hasException);
-        return histogramSlot.getSlotTime();
     }
 
     private List<TransactionId> recursiveCallFilter(List<TransactionId> transactionIdList) {
